move delete expired challenges to token api as /challenges/expired
