+++ privacyidea/api/token.py
@@ -83,7 +83,7 @@ from privacyidea.lib.importotp import (parseOATHcsv, parseSafeNetXML,
                                        parseYubicoCSV, parsePSKCdata, GPGImport)
 import logging
 from privacyidea.lib.policy import ACTION
-from privacyidea.lib.challenge import get_challenges_paginate
+from privacyidea.lib.challenge import get_challenges_paginate, cleanup_expired_challenges
 from privacyidea.api.lib.prepolicy import (prepolicy, check_base_action, check_token_action,
                                            check_token_init, check_token_upload,
                                            check_max_token_user,
@@ -382,6 +382,36 @@ def get_challenges_api(serial=None):
                                          sortdir=sdir, page=page, psize=psize)
     g.audit_object.log({"success": True})
     return send_result(challenges)
+
+@token_blueprint.route("/challenges/expired", methods=['DELETE'])
+@admin_required
+@log_with(log)
+def delete_challenge_cache_api():
+    """
+    Delete expired entries in the challenge table.
+
+    :>json bool status: Status of the request
+    :reqheader PI-Authorization: The authorization token
+
+    **Example response**:
+
+    .. sourcecode:: http
+
+        HTTP/1.1 200 OK
+        Content-Type: application/json
+
+        {
+          "id": 1,
+          "jsonrpc": "2.0",
+          "result": {
+            "status": true
+          },
+          "version": "privacyIDEA unknown"
+        }
+    """
+    row_count = cleanup_expired_challenges(chunksize=None, age=None)
+    g.audit_object.log({"success": True, "info": f"Deleted {row_count} entries from challenges"})
+    return send_result({"status": True})
 
 
 @token_blueprint.route('/', methods=['GET'])
