+++ privacyidea/static/components/token/views/token.challenges.html
@@ -2,7 +2,7 @@
     {{ challengedata.count }}
 </p>
 <button type="button"
-        ng-click="deleteChallengeCache()"
+        ng-click="deleteExpiredChallenges()"
         class="btn btn-primary"
         translate>
 Delete Expired Challenges
